only check host anchored filters when necessary
A trie uses too much memory and takes too much time to create for host
anchored filters. Splitting the filters into groups based on the last
two letters of their host, and then checking only the filters with the
same domain ending as the domain ending for the request, makes checking
filters significantly faster, and doesn’t take as long to parse.