Fixes #311
Replaces each tab element's `click` event listener with separate `mousedown`, `mousemove` and `mouseup` event listeners which determine whether or not the user dragged the mouse during the click.

If a drag was detected on the tab element, a flag is toggled and the action for the click won't happen.

This is the same solution mentioned in [this](http://stackoverflow.com/questions/6042202/how-to-distinguish-mouse-click-and-drag) StackOverflow post.