move taskContainer and closeButton creation to obj
Moved these two functions to the TaskOverlayBuilder namespace
Now that all creation-connected functions (except one) are split into
pieces and grouped in a namespace it should be very easy to extract
overlay building from using it into separate files. This greatly reduces
complexity and improves readability.
It also allows much easier creation of extensions to the taskOverlay. A
few examples are keyboard-oriented control of the taskOverlay,
fuzzy-search of tabs and more control buttons for tabs.