Convert the array-like `tasks` object to a TaskList instance
Makes it easier to reason about, because now it is not a mix of an array
and an object.
Instead, the array is kept as a reference inside TaskList,
and the global tasks variable is an instance of it.

Avoids the need of iterating a prototype object and assigning its props
to an array.