add settings section for clearing browser history
Minimal version of the settings pane to clear browsing history.

Sadly, the functionality is not there:

1. First attempt was to use `places.deleteAllHistory()` but it's not available in this scope and I can't use require/import to fetch it

2. Tried using directly postMessage as places internally does but that postMessage within places is directed to the particular worker so this didn't work either

Also used inline styles which should be migrated to some stylesheet, but just wanted to put a quick PoC together to ask for pointers when it comes to actually clearing the DB of its stored URLs.